pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "zipxml"
version = "1.0"
dependencies = [
    "typer-cli"
]
